requirements-lint.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=lint --output-file=requirements-lint.txt pyproject.toml
#
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
black==23.7.0
    # via turtget (pyproject.toml)
click==8.1.6
    # via black
comm==0.1.3
    # via ipykernel
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via ipython
executing==1.2.0
    # via stack-data
flake8==6.1.0
    # via turtget (pyproject.toml)
ipykernel==6.25.0
    # via ipywidgets
ipython==8.14.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.7
    # via turtget (pyproject.toml)
jedi==0.19.0
    # via ipython
jupyter-client==8.3.0
    # via ipykernel
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.8
    # via ipywidgets
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
nest-asyncio==1.5.7
    # via ipykernel
packaging==23.1
    # via
    #   black
    #   ipykernel
parso==0.8.3
    # via jedi
pathspec==0.11.2
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.0
    # via turtget (pyproject.toml)
platformdirs==3.10.0
    # via
    #   black
    #   jupyter-core
prompt-toolkit==3.0.39
    # via ipython
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.0
    # via flake8
pyflakes==3.1.0
    # via flake8
pygments==2.15.1
    # via ipython
python-dateutil==2.8.2
    # via jupyter-client
pyzmq==25.1.0
    # via
    #   ipykernel
    #   jupyter-client
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
stack-data==0.6.2
    # via ipython
tomli==2.0.1
    # via black
tornado==6.3.2
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
wcwidth==0.2.6
    # via prompt-toolkit
widgetsnbextension==4.0.8
    # via ipywidgets
